# Conflicts:
#	Makefile
#	src/main/config/config.c
#	src/main/drivers/accgyro_mpu.c
#	src/main/drivers/accgyro_mpu6500.c
#	src/main/drivers/gyro_sync.c
#	src/main/flight/imu.c
#	src/main/io/serial_1wire.c
#	src/main/io/serial_cli.c
#	src/main/sensors/gyro.c
#	src/main/sensors/gyro.h

+++ src/main/drivers/accgyro_l3gd20.c
@@ -86,7 +86,7 @@ static void l3gd20SpiInit(SPI_TypeDef *SPIx)
     spiSetDivisor(L3GD20_SPI, SPI_STANDARD_CLOCK);
 }
 
-void l3gd20GyroInit(uint16_t lpf)
+void l3gd20GyroInit(uint8_t lpf)
 {
     UNUSED(lpf); // FIXME use it!
